[VOGRE-37] Code cleanup
