--- concepts/Templates/login/test.html
@@ -1,203 +1,0 @@
-{% extends "account/base.html" %}
-
-{% load i18n %}
-
-{% block head_links %}
-<title>{% trans "Sign In to ConceptPower" %}</title>
-<style>
-    /* Custom styles for ConceptPower login page */
-    .conceptpower-card {
-        border-radius: 10px;
-        box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.1);
-    }
-
-    .conceptpower-header {
-        background: linear-gradient(to right, #007bff, #0056b3);
-        color: white;
-        border-radius: 10px 10px 0 0;
-    }
-
-    .conceptpower-container {
-        margin-top: 50px;
-    }
-
-    .form-group input {
-        border-radius: 5px;
-    }
-
-    .btn-custom {
-        background-color: #007bff;
-        border: none;
-        border-radius: 5px;
-        padding: 10px;
-        font-size: 16px;
-        transition: all 0.3s ease;
-    }
-
-    .btn-custom:hover {
-        background-color: #0056b3;
-    }
-
-    .info-text {
-        font-size: 14px;
-        color: #6c757d;
-    }
-</style>
-{% endblock %}
-
-{% block head_title %}{% trans "ConceptPower Sign In" %}{% endblock %}
-
-{% block content %}
-<div class="container conceptpower-container">
-    {% block messages %}
-    {% if messages %}
-    <div>
-        <ul>
-        {% for message in messages %}
-        <div class="alert alert-warning alert-dismissible fade show" role="alert">
-            <strong>{{ message }}</strong>
-            <button type="button" class="close" data-dismiss="alert" aria-label="Close">
-                <span aria-hidden="true">&times;</span>
-            </button>
-        </div>
-        {% endfor %}
-        </ul>
-    </div>
-    {% endif %}
-    {% endblock %}
-
-    <div class="row justify-content-center">
-        <div class="col-md-6 col-lg-5">
-            <div class="card conceptpower-card">
-                <div class="card-header text-center conceptpower-header">
-                    <h3 class="h3" id="head_banner">{% trans "ConceptPower Sign In" %}</h3>
-                </div>
-                <div class="card-body">
-                    <p class="text-center info-text">
-                        {% trans "Enter your ConceptPower credentials to continue." %}
-                    </p>
-
-                    <form class="login" method="POST" action="{% url 'conceptpower_login' %}">
-                        {% csrf_token %}
-
-                        <!-- If using Django Form -->
-                        {{ form.as_p }}
-
-                        <!-- If manually defining fields -->
-                        <div class="form-group mb-3">
-                            <label for="username">{% trans "ConceptPower Username" %}</label>
-                            <input type="text" name="username" class="form-control" id="username" required placeholder="{% trans 'Enter your ConceptPower username' %}">
-                        </div>
-
-                        <div class="form-group mb-3">
-                            <label for="password">{% trans "ConceptPower Password" %}</label>
-                            <input type="password" name="password" class="form-control" id="password" required placeholder="{% trans 'Enter your ConceptPower password' %}">
-                        </div>
-
-                        {% if redirect_field_value %}
-                        <input type="hidden" name="{{ redirect_field_name }}" value="{{ redirect_field_value }}" />
-                        {% endif %}
-
-                        <div class="form-group text-center">
-                            <a class="btn btn-link" href="{% url 'account_reset_password' %}">
-                                {% trans "Forgot Password?" %}
-                            </a>
-                        </div>
-
-                        <div class="d-grid gap-2">
-                            <button class="mt-3 btn btn-custom btn-block" type="submit">
-                                {% trans "Sign In" %}
-                            </button>
-                        </div>
-                    </form>
-                </div>
-            </div>
-
-            <!-- Section for Adding/Updating ConceptPower Credentials -->
-            <div class="mt-4 text-center">
-                <h5>{% trans "Don't have ConceptPower credentials?" %}</h5>
-                <p class="info-text">{% trans "Add your credentials below to access ConceptPower features." %}</p>
-
-                <button class="btn btn-secondary" data-toggle="modal" data-target="#conceptpowerModal">
-                    {% trans "Add / Update Credentials" %}
-                </button>
-            </div>
-        </div>
-    </div>
-</div>
-
-<!-- Modal for Adding / Updating ConceptPower Credentials -->
-<div class="modal fade" id="conceptpowerModal" tabindex="-1" role="dialog" aria-labelledby="conceptpowerModalLabel" aria-hidden="true">
-    <div class="modal-dialog" role="document">
-        <div class="modal-content">
-            <div class="modal-header">
-                <h5 class="modal-title" id="conceptpowerModalLabel">{% trans "Manage ConceptPower Credentials" %}</h5>
-                <button type="button" class="close" data-dismiss="modal" aria-label="Close">
-                    <span aria-hidden="true">&times;</span>
-                </button>
-            </div>
-            <div class="modal-body">
-                <form id="conceptpowerForm">
-                    <div class="form-group">
-                        <label for="new_username">{% trans "ConceptPower Username" %}</label>
-                        <input type="text" class="form-control" id="new_username" required>
-                    </div>
-
-                    <div class="form-group">
-                        <label for="new_password">{% trans "ConceptPower Password" %}</label>
-                        <input type="password" class="form-control" id="new_password" required>
-                    </div>
-
-                    <button type="submit" class="btn btn-primary btn-block mt-3">{% trans "Save Credentials" %}</button>
-                </form>
-            </div>
-        </div>
-    </div>
-</div>
-
-<!-- JavaScript to Handle Credential Submission -->
-<script>
-    document.getElementById("conceptpowerForm").addEventListener("submit", async function(event) {
-        event.preventDefault();
-
-        const newUsername = document.getElementById("new_username").value;
-        const newPassword = document.getElementById("new_password").value;
-
-        const response = await fetch("{% url 'conceptpower_login' %}", {
-            method: "POST",
-            headers: {
-                "Content-Type": "application/json",
-                "X-CSRFToken": getCookie("csrftoken"), // Ensure CSRF token is included
-            },
-            body: JSON.stringify({
-                username: newUsername,
-                password: newPassword
-            })
-        });
-
-        const result = await response.json();
-        if (response.ok && result.success) {
-            alert("ConceptPower credentials updated successfully!");
-            window.location.reload(); // Refresh to show updated state
-        } else {
-            alert("Error: " + (result.message || "Failed to update credentials."));
-        }
-    });
-
-    // Helper function to get CSRF token
-    function getCookie(name) {
-        let cookieValue = null;
-        if (document.cookie && document.cookie !== '') {
-            const cookies = document.cookie.split(';');
-            for (let i = 0; i < cookies.length; i++) {
-                const cookie = cookies[i].trim();
-                if (cookie.substring(0, name.length + 1) === (name + '=')) {
-                    cookieValue = decodeURIComponent(cookie.substring(name.length + 1));
-                    break;
-                }
-            }
-        }
-        return cookieValue;
-    }
-</script>
-{% endblock %}
