+++ external_accounts/admin.py
@@ -3,3 +3,4 @@ from external_accounts.models import *
 
 # Register your models here.
 admin.site.register(CitesphereAccount)
+admin.site.register(ConceptpowerAccount)
